STS2024/code/requirements.txt
acvl_utils==0.2
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
batchgenerators==0.25
batchgeneratorsv2==0.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
connected-components-3d==3.18.0
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dicom2nifti==2.4.11
docopt==0.6.2
dynamic_network_architectures==0.3.1
einops==0.8.0
et-xmlfile==1.1.0
executing==2.1.0
fastjsonschema==2.20.0
fft-conv-pytorch==1.2.0
filelock==3.15.4
filetype==1.2.0
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
graphviz==0.20.3
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
imagecodecs==2024.6.1
imageio==2.34.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lazy_loader==0.4
linecache2==1.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1.post1
matplotlib-inline==0.1.7
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
nibabel==5.2.1
nnunetv2==2.5.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pip==24.2
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycparser==2.22
pydicom==2.4.4
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-gdcm==3.0.24.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.40
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
setuptools==72.1.0
SimpleITK==2.3.1
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.0
threadpoolctl==3.5.0
tifffile==2024.7.24
tinycss2==1.3.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traceback2==1.4.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.79
ultralytics-thop==2.0.5
unittest2==1.1.0
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.44.0
win-inet-pton==1.1.0
yacs==0.1.8
yarg==0.1.9
zstandard==0.23.0
